OS X platform implementation
