+++ projects/openrct2.vcxproj
@@ -85,8 +85,8 @@
     <ClCompile Include="..\src\peep\peep.c" />
     <ClCompile Include="..\src\peep\staff.c" />
     <ClCompile Include="..\src\platform\linux.c" />
+    <ClCompile Include="..\src\platform\posix.c" />
     <ClCompile Include="..\src\platform\shared.c" />
-    <ClCompile Include="..\src\platform\unix.c" />
     <ClCompile Include="..\src\platform\osx.c" />
     <ClCompile Include="..\src\platform\windows.c" />
     <ClCompile Include="..\src\rct1.c" />
